fix: إصلاح أخطاء الحاسبات
✅ الإصلاحات:
- حذف _buildResultCard المكرر في roi_calculator_page
- تغيير _convert() إلى _calculate() في savings_calculator_page
- تحويل bathroomCost و kitchenCost إلى double في finishing_cost_calculator_page
- جميع الأخطاء محلولة

+++ lib/finishing_cost_calculator_page.dart
@@ -58,8 +58,8 @@ class _FinishingCostCalculatorPageState extends State<FinishingCostCalculatorPag
       _breakdown = {
         'الأرضيات': floorCost,
         'الدهان': paintCost,
-        'الحمامات': bathroomCost,
-        'المطبخ': kitchenCost,
+        'الحمامات': bathroomCost.toDouble(),
+        'المطبخ': kitchenCost.toDouble(),
         'الكهرباء': electricCost,
         'السباكة': plumbingCost,
       };
